backend/requirements.txt
fastapi
uvicorn[standard]
requests
google-genai
python-dotenv
Pillow
PyPDF2
